Topic 6. Model Inference/inference-python/requirements.txt
opencv-python==4.5.5.62
pydantic==1.9.0
requests==2.27.1
Pillow==8.0.1
marshmallow==3.10.0
Flask==2.2.2
itsdangerous==2.0.1
pandas==1.3.5
matplotlib>=3.3
seaborn==0.12.2
